refactoring: More refactoring on the LineSelectionScreenWidget
